Move LSP requests to individual modules
